requirements.txt
Package                  Version
------------------------ -----------
accelerate               1.10.0
aioboto3                 15.1.0
aiobotocore              2.24.0
aiofiles                 24.1.0
aiohappyeyeballs         2.6.1
aiohttp                  3.12.15
aioitertools             0.12.0
aiosignal                1.4.0
annotated-types          0.7.0
anyio                    4.10.0
async-timeout            4.0.3
attrs                    25.3.0
beautifulsoup4           4.13.4
boto3                    1.39.11
botocore                 1.39.11
certifi                  2025.8.3
charset-normalizer       3.4.3
click                    8.2.1
cssselect                1.3.0
dataclasses-json         0.6.7
deeplake                 4.2.14
dill                     0.4.0
distro                   1.9.0
exceptiongroup           1.3.0
feedfinder2              0.0.4
feedparser               6.0.11
filelock                 3.19.1
frozenlist               1.7.0
fsspec                   2025.7.0
gpt4all                  2.8.2
greenlet                 3.2.4
h11                      0.16.0
hf-xet                   1.1.8
httpcore                 1.0.9
httpx                    0.28.1
httpx-sse                0.4.1
huggingface-hub          0.34.4
humbug                   0.3.2
idna                     3.10
jieba3k                  0.35.1
Jinja2                   3.1.6
jiter                    0.10.0
jmespath                 1.0.1
joblib                   1.5.1
jsonpatch                1.33
jsonpointer              3.0.0
langchain                0.3.27
langchain-community      0.3.27
langchain-core           0.3.74
langchain-deeplake       0.1.0
langchain-huggingface    0.3.1
langchain-openai         0.3.30
langchain-text-splitters 0.3.9
langsmith                0.4.14
lxml                     6.0.0
lxml_html_clean          0.4.2
lz4                      4.4.4
MarkupSafe               3.0.2
marshmallow              3.26.1
mpmath                   1.3.0
multidict                6.6.4
multiprocess             0.70.18
mypy_extensions          1.1.0
nest-asyncio             1.6.0
networkx                 3.4.2
newspaper3k              0.2.8
nltk                     3.9.1
numpy                    1.26.4
nvidia-cublas-cu12       12.8.4.1
nvidia-cuda-cupti-cu12   12.8.90
nvidia-cuda-nvrtc-cu12   12.8.93
nvidia-cuda-runtime-cu12 12.8.90
nvidia-cudnn-cu12        9.10.2.21
nvidia-cufft-cu12        11.3.3.83
nvidia-cufile-cu12       1.13.1.3
nvidia-curand-cu12       10.3.9.90
nvidia-cusolver-cu12     11.7.3.90
nvidia-cusparse-cu12     12.5.8.93
nvidia-cusparselt-cu12   0.7.1
nvidia-nccl-cu12         2.27.3
nvidia-nvjitlink-cu12    12.8.93
nvidia-nvtx-cu12         12.8.90
openai                   1.99.9
orjson                   3.11.2
outcome                  1.3.0.post0
packaging                25.0
pathos                   0.3.4
pillow                   10.4.0
pip                      24.3.1
pox                      0.3.6
ppft                     1.7.7
propcache                0.3.2
psutil                   7.0.0
pydantic                 2.11.7
pydantic_core            2.33.2
pydantic-settings        2.10.1
PyJWT                    2.10.1
pypdf                    6.0.0
PySocks                  1.7.1
python-dateutil          2.9.0.post0
python-dotenv            1.1.1
PyYAML                   6.0.2
regex                    2025.7.34
requests                 2.32.4
requests-file            2.1.0
requests-toolbelt        1.0.0
s3transfer               0.13.1
safetensors              0.6.2
selenium                 4.35.0
sentencepiece            0.2.1
setuptools               44.1.1
sgmllib3k                1.0.0
six                      1.16.0
sniffio                  1.3.1
sortedcontainers         2.4.0
soupsieve                2.7
SQLAlchemy               2.0.43
sympy                    1.14.0
tenacity                 9.1.2
tiktoken                 0.11.0
tinysegmenter            0.3
tldextract               5.3.0
tokenizers               0.21.4
torch                    2.8.0
tqdm                     4.67.1
transformers             4.55.2
trio                     0.30.0
trio-websocket           0.12.2
triton                   3.4.0
typing_extensions        4.14.1
typing-inspect           0.9.0
typing-inspection        0.4.1
urllib3                  2.5.0
websocket-client         1.8.0
wrapt                    1.17.3
wsproto                  1.2.0
yarl                     1.20.1
zstandard                0.24.0
